+++ Craftsman/Builders/CoreExceptionsBuilder.cs
@@ -21,6 +21,23 @@
 
             CreateApiException(solutionDirectory, projectBaseName);
             CreateValidationException(solutionDirectory, projectBaseName);
+            CreateConflictException(solutionDirectory, projectBaseName);
+        }
+
+        public static void CreateConflictException(string solutionDirectory, string projectBaseName)
+        {
+            var classPath = ClassPathHelper.CoreExceptionClassPath(solutionDirectory, $"ConflictException.cs", projectBaseName);
+
+            if (!Directory.Exists(classPath.ClassDirectory))
+                Directory.CreateDirectory(classPath.ClassDirectory);
+
+            if (File.Exists(classPath.FullClassPath))
+                throw new FileAlreadyExistsException(classPath.FullClassPath);
+
+            using FileStream fs = File.Create(classPath.FullClassPath);
+            var data = "";
+            data = GetConflictExceptionFileText(classPath.ClassNamespace);
+            fs.Write(Encoding.UTF8.GetBytes(data));
         }
 
         public static void CreateApiException(string solutionDirectory, string projectBaseName)
@@ -55,6 +72,27 @@
         public ApiException(string message) : base(message) {{ }}
 
         public ApiException(string message, params object[] args)
+            : base(string.Format(CultureInfo.CurrentCulture, message, args))
+        {{
+        }}
+    }}
+}}";
+        }
+
+        public static string GetConflictExceptionFileText(string classNamespace)
+        {
+            return @$"namespace {classNamespace}
+{{
+    using System;
+    using System.Globalization;
+
+    public class ConflictException : Exception
+    {{
+        public ConflictException() : base() {{ }}
+
+        public ConflictException(string message) : base(message) {{ }}
+
+        public ConflictException(string message, params object[] args)
             : base(string.Format(CultureInfo.CurrentCulture, message, args))
         {{
         }}