+++ Craftsman/Builders/ErrorHandlerMiddlewareBuilder.cs
@@ -69,19 +69,21 @@
 
                 switch (error)
             {{
-                    case ApiException e:
+                    case ConflictException:
+                        response.StatusCode = (int)HttpStatusCode.Conflict;
+                        break;
+
+                    case ApiException:
                         // custom application error
                         response.StatusCode = (int)HttpStatusCode.BadRequest;
                         break;
 
                     case ValidationException e:
-                        // custom application error
                         response.StatusCode = (int)HttpStatusCode.BadRequest;
                         responseModel.Errors = e.Errors;
                         break;
 
-                    case KeyNotFoundException e:
-                        // not found error
+                    case KeyNotFoundException:
                         response.StatusCode = (int)HttpStatusCode.NotFound;
                         break;