+++ Craftsman/Helpers/Extensions.cs
@@ -16,6 +16,7 @@
             // Return char and concat substring.
             return char.ToUpper(value[0]) + value.Substring(1);
         }
+
         public static string LowercaseFirstLetter(this string value)
         {
             // Check for empty string.
@@ -34,6 +35,11 @@
                 input = input.Substring(0, index) + input.Substring(index + Environment.NewLine.Length);
 
             return input;
+        }
+
+        public static bool IsGuidPropertyType(this string input)
+        {
+            return input.Equals("guid", StringComparison.InvariantCultureIgnoreCase);
         }
     }
 }